Main logo font selection
On the 22nd of April, the client suggested that the font of the logo could be changed to better reflect the meaning of 'Digit Dash.' This photo includes some of the fonts which I combined in search for the perfect logo. At the end, I resorted to using the same font as the tutorial screens, yet with the word 'Dash' containing dashes for a speed-like effect.