Improved map curves, static level preview screen
To ensure that the bezier curves were different between each level, I had to create static curve templates which I store within the levels list. Furthermore, I began development on the level preview screen, yet the only current functionality is for the user to navigate back to the levels page.