Start page, game timer and tutorial overlay added + bug fixes
Being one of the biggest commits in the project, this daily update contains a large number of improvements and new features to the game. Now users are greeted with a welcome screen, asking whether they would like a tutorial or not. For now, the tutorial only has one screen for testing purposes, after a lot of trial and error, I was able to make the overlay with empty spaces work so that the user can see highlited screen components. The gameplay now works completely, with or without a timer. At the end of the game, the user is presented with the number of questions they got correct. Furthermore, I added a save game data functionality which ensures that the user's progress is saved even if they exit the app. Now when the user gets the target number of correct questions for a level, the next one is unlocked for them! I also added some other cosmetic changes and a simple development functionality which allows me to remove saved game data for debugging purposes.