Added in-game soundtracks, finalised tutorial
I recorded a series of soundtracks with my own voice using Audacity, exported them as mp3 files and created a few triggers during gameplay for the sounds to play. When the user gets a question correct or incorrect, a random sound (from the ones I created) will play; the same occurs when the game timer runs out. Furthermore, the tutorial now includes 4 different screens, which lead to an actual game that can be played by the user. I've also been working on improving the implemented designs in some screens, such as greying out trophies when the user hasn't earnt them.