Improved game stability, repetition probability study
A series of minor bugs were fixed whilst carrying out the application's testing phase. This included an index error for the map curves when first installing the application, an issue with the calculation of stars when a level is finished, ensuring that the user can only enter a maximum of 3 digits as an answer and a boolean logic problem with the tutorial overlay. Furthermore, I also ensured that the AlertDialog for when a user doesn't enter a value for their answer is shown and works correctly.