No repeated questions, different sounds
Apart from a bug fix with the number of stars and whether the next level was already unlocked or not on the game end screen, I also ensured that no questions were repeated throughout a round by adding each generated question to a list and then checking whether a new question was already generated previously. Finally, I also ensured that the game doesn't play the same sound twice in a row.