In-game menu and star rating fixes, asset cleanup
After adding the final optimisations and fixing general bugs in the application, I came across some issues regarding the newly implemented in-game menu as well as the end-screen star display. To fix an issue with the game not restarting correctly when the user chooses to do so in the in-game menu, I ensured that the game's timer is completely stopped before the first question of the round is displayed again. Furthermore, I also ensured that the star score calculation allowed the user to obtain 3 stars if the user's score is above the target (not the maximum score) for the level. Finally, I cross-checked with the assets table in the main assignment document and ensured that any unused assets were deleted on both the table and the actual solution's file directory.