Screen size optimisation, progress update
After testing out the application on the Samsung Galaxy A70, only two small problems arose; the tutorial tick button was off it's intended position by around 20 pixels on the y axis and the tick text on the main screen was black instead of white. These issues are patched in this commit whilst ensuring that any progress outside of GitHub is also updated according to the tasks completed over the last two days.